requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selene==2.0.0b6
requests
voluptuous
pytest_voluptuous
faker
python-dotenv==0.20.0
